Extract the precompute library from the Modal app
The precompute app had outgrown the self-contained-ephemeral-app shape of
smoke_app/prewarm_app: it carried real library logic. Function bodies,
planning, manifest assembly, and the local-entrypoint orchestration move
to policyengine_simulation_executor.precompute, mirroring the deployed
app's split (run_simulation -> run_simulation_impl); the app file keeps
only image, declarations, and thin wrappers. The orchestration takes its
five Modal function handles as parameters, so the whole flow — miss-only
spawning, the determinism-gate probe, the abort-before-publish path, and
the MANIFEST_DIGEST= stdout contract the deploy job will parse — now has
direct test coverage with plain fakes. The fake-modal scaffolding the app
tests still need is shared from fixtures/fake_modal.py instead of being
copied per test file, and the dataset builder now goes through
_country_module like the rest of the runtime.
