+++ projects/policyengine-simulation-executor/tests/test_modal_bundle_image.py
@@ -2,76 +2,8 @@ import importlib
 import sys
 import tomllib
 from pathlib import Path
-from types import ModuleType
 
-class FakeImage:
-    def __init__(self):
-        self.calls = []
-
-    @classmethod
-    def debian_slim(cls, python_version):
-        image = cls()
-        image.calls.append(("debian_slim", python_version))
-        return image
-
-    def pip_install(self, *packages):
-        self.calls.append(("pip_install", packages))
-        return self
-
-    def pip_install_from_requirements(self, requirements_txt, **kwargs):
-        self.calls.append(
-            ("pip_install_from_requirements", requirements_txt, kwargs)
-        )
-        return self
-
-    def uv_sync(self, uv_project_dir="./", **kwargs):
-        self.calls.append(("uv_sync", uv_project_dir, kwargs))
-        return self
-
-    def run_commands(self, *commands, **kwargs):
-        self.calls.append(("run_commands", commands, kwargs))
-        return self
-
-    def env(self, env):
-        self.calls.append(("env", env))
-        return self
-
-    def add_local_python_source(self, *args, **kwargs):
-        self.calls.append(("add_local_python_source", args, kwargs))
-        return self
-
-    def run_function(self, function, **kwargs):
-        self.calls.append(("run_function", function.__name__, kwargs))
-        return self
-
-
-class FakeSecret:
-    @staticmethod
-    def from_name(*args, **kwargs):
-        return {"args": args, "kwargs": kwargs}
-
-
-class FakeApp:
-    def __init__(self, name):
-        self.name = name
-        self.function_calls = []
-
-    def function(self, **kwargs):
-        def decorator(function):
-            self.function_calls.append((function.__name__, kwargs))
-            return function
-
-        return decorator
-
-
-def install_fake_modal(monkeypatch):
-    modal = ModuleType("modal")
-    modal.Image = FakeImage
-    modal.Secret = FakeSecret
-    modal.App = FakeApp
-    modal.is_local = lambda: True
-    modal.asgi_app = lambda: lambda function: function
-    monkeypatch.setitem(sys.modules, "modal", modal)
+from fixtures.fake_modal import install_fake_modal
 
 
 def test_modal_image_uses_policyengine_bundle_install(monkeypatch):
@@ -213,9 +145,7 @@ def test_app_module_imports_at_container_entrypoint_path(monkeypatch):
     monkeypatch.setenv("POLICYENGINE_UK_VERSION", "2.90.0")
     sys.modules.pop("src.modal.app", None)
 
-    source_path = (
-        Path(__file__).resolve().parents[1] / "src" / "modal" / "app.py"
-    )
+    source_path = Path(__file__).resolve().parents[1] / "src" / "modal" / "app.py"
     code = compile(source_path.read_text(), "/root/app.py", "exec")
     spec = importlib.util.spec_from_loader(
         "container_entrypoint_app", loader=None, origin="/root/app.py"
